Spring boot

+++ src/main/java/com/ust/pos/user/UserController.java
@@ -3,7 +3,6 @@ package com.ust.pos.user;
 import com.ust.pos.dto.UserDto;
 import com.ust.pos.role.service.RoleService;
 import com.ust.pos.user.service.UserService;
-import org.jspecify.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -49,17 +48,16 @@ public class UserController {
 
     @GetMapping("/delete")
     public String delete(Model model, @RequestParam String username) {
-
-        @Nullable Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String loggedInUser = authentication.getName();
-
-        userService.delete(username);
-
-        if (loggedInUser.equals(username)) {
-            SecurityContextHolder.clearContext();
-            return "redirect:/login";
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null) {
+            String loggedInUser = authentication.getName();
+            userService.delete(username);
+            if (loggedInUser.equals(username)) {
+                SecurityContextHolder.clearContext();
+                return "redirect:/logout";
+            }
+            return REDIRECT_USER_LIST;
         }
-
-        return "redirect:/user/list";
+        return null;
     }
 }
